// src/components/sidebar/workflows/types.ts
export interface HistoryEntry {
  type: 'state_add' | 'state_delete' | 'state_move' | 'transition_add' | 'transition_delete'
  data: any
}

export interface ClipboardData {
  type: 'state' | 'transition'
  data: any
}